fix(ChargePrices): Fetching now the correct prices
The Chargeprice api requires that "charge_points.plug" Field inside the JSON
Request Body has to be lowercase instead of uppercase.

I also disabled the "hyundai inoque" vehicle because can override
other cached prices. We need to find a better solution for that.
The Vehicle table got also a switch to disable some vehicles.
Temporally the "hyundai inoque" tariff will not be retrieved from the database.

Refs: #21