feat:(Image Cache)
This feature can be now tested.
DB Migration is done. The "tarif" table has a new internal_name field.
Renaming all internal "tarif" names done.
The server will watch for a images folder and will keep the db up to date.
This feature is available in v2 of the API.
Three more fields have been added to the cards JSON: image, tarif_url and blocking_fee_start.
The image field is an URL which can be null for future prove.
In Addition to that an /images/card/{endpoint} has be added.